refactor: new mobile menu
