+++ src/components/shared/mobile-menu/mobile-menu.jsx
@@ -7,9 +7,11 @@ import Button from 'components/shared/button';
 import Link from 'components/shared/link';
 import LINKS from 'constants/links';
 import MENUS from 'constants/menus';
+import useScrollStatus from 'hooks/use-scroll-status';
 import ChevronIcon from 'icons/chevron-small.inline.svg';
 
 import SubMenu from './sub-menu';
+import TabletMenu from './tablet-menu';
 
 const ANIMATION_DURATION = 0.2;
 
@@ -32,10 +34,11 @@ const variants = {
 };
 
 const MobileMenu = ({ isOpen }) => {
-  const [currentMenu, setCurrentMenu] = useState(null);
+  const [openMenu, setOpenMenu] = useState(null);
+  const { isScrolledToBottom, hasScroll, handleScroll } = useScrollStatus();
 
-  const handleOpenMenu = (menuItems) => {
-    setCurrentMenu((current) => (current?.label === menuItems?.label ? null : menuItems));
+  const handleOpenMenu = (items) => {
+    setOpenMenu((current) => (current?.label === items?.label ? null : items));
   };
 
   return (
@@ -43,19 +46,27 @@ const MobileMenu = ({ isOpen }) => {
       <AnimatePresence>
         {isOpen && (
           <m.div
-            className="safe-paddings pointer-events-none fixed inset-0 hidden w-full flex-col justify-between overflow-x-hidden overflow-y-scroll pt-16 sm:flex"
+            className="safe-paddings pointer-events-none fixed inset-0 hidden w-full flex-col justify-between overflow-x-hidden overflow-y-scroll pt-16 md:flex"
             initial="hidden"
             animate="visible"
             exit="hidden"
             variants={variants}
           >
-            <nav className="pointer-events-auto flex h-full w-full overflow-x-hidden overflow-y-scroll bg-black">
-              <ul className="relative flex h-full w-full flex-col border-t border-t-[#1F1F1F] bg-black">
-                {MENUS.header.map(({ text, menuItems, to, target }, index) => (
-                  <li className="border-b border-b-gray-2 px-5 last:border-b-0" key={index}>
+            <nav
+              className={clsx(
+                'pointer-events-auto relative flex h-full w-full overflow-x-hidden overflow-y-scroll bg-black',
+                hasScroll &&
+                  !isScrolledToBottom &&
+                  'after:pointer-events-none after:fixed after:inset-x-4 after:bottom-24 after:top-16 after:z-50 after:bg-[linear-gradient(180deg,#05050B00_86.18%,#05050B_100%)] sm:after:bottom-[142px]'
+              )}
+              onScroll={handleScroll}
+            >
+              <ul className="relative flex h-full w-full flex-col border-t border-t-[#1F1F1F] px-8 sm:px-5">
+                {MENUS.header.map(({ text, to, target, menuItems }, index) => (
+                  <li className="border-b border-b-[#1F1F1F] last:border-b-0" key={index}>
                     <Link
                       className={clsx(
-                        'flex h-[45px] w-full items-center justify-between text-left text-base leading-none'
+                        'flex h-[53px] w-full items-center justify-between text-left text-lg leading-none sm:h-[45px] sm:text-base'
                       )}
                       tag={to ? null : 'button'}
                       to={to}
@@ -63,25 +74,31 @@ const MobileMenu = ({ isOpen }) => {
                       target={target}
                       onClick={to ? null : () => handleOpenMenu(menuItems)}
                     >
-                      <span>{text}</span>
-                      {menuItems && <ChevronIcon className="h-3.5 w-3.5 -rotate-90" />}
+                      {text}
+                      {menuItems && (
+                        <ChevronIcon
+                          className={clsx(
+                            'size-3.5 -rotate-90 transition-transform duration-200',
+                            openMenu?.label === menuItems?.label ? 'rotate-0' : ''
+                          )}
+                        />
+                      )}
                     </Link>
+                    <TabletMenu openMenu={openMenu} label={menuItems?.label} />
                   </li>
                 ))}
               </ul>
-              <SubMenu currentMenu={currentMenu} handleOpenMenu={handleOpenMenu} />
+              <SubMenu currentMenu={openMenu} handleOpenMenu={handleOpenMenu} />
             </nav>
 
             <div className="pointer-events-auto bg-black">
-              <div className="container px-5">
-                <div className="flex w-full flex-col justify-between gap-3.5 py-6">
-                  <Button className="w-full" size="xs" theme="white-outline">
-                    Login
-                  </Button>
-                  <Button className="w-full" size="xs" theme="white-filled" {...LINKS.getStarted}>
-                    Get Started
-                  </Button>
-                </div>
+              <div className="container flex justify-between gap-x-4 gap-y-3.5 py-7 sm:px-5 sm:py-6 xs:flex-col">
+                <Button className="w-full" size="xs" theme="white-outline">
+                  Login
+                </Button>
+                <Button className="w-full" size="xs" theme="white-filled" {...LINKS.getStarted}>
+                  Get Started
+                </Button>
               </div>
             </div>
           </m.div>
